# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Laporan Harian - Daily Work Report System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /po:
    get:
      operationId: listPo
      tags: [po]
      summary: List PO/Projects with filters
      parameters:
        - name: month
          in: query
          schema: { type: integer }
        - name: year
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema: { type: string }
        - name: departmentId
          in: query
          schema: { type: integer }
        - name: picUserId
          in: query
          schema: { type: integer }
        - name: customer
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of PO
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PoItem"
    post:
      operationId: createPo
      tags: [po]
      summary: Create a new PO/Project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PoInput"
      responses:
        "201":
          description: Created PO
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoItem"
  /po/summary:
    get:
      operationId: getPoSummary
      tags: [po]
      summary: Get monthly PO summary stats
      parameters:
        - name: month
          in: query
          schema: { type: integer }
        - name: year
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: PO summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoSummary"
  /po/{id}:
    get:
      operationId: getPo
      tags: [po]
      summary: Get PO detail
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: PO detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoItem"
    patch:
      operationId: updatePo
      tags: [po]
      summary: Update a PO/Project
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PoUpdate"
      responses:
        "200":
          description: Updated PO
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoItem"
  /po/{id}/close:
    post:
      operationId: closePo
      tags: [po]
      summary: Mark PO as closed/completed
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Closed PO
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoItem"
...
components:
  schemas:
...
    PoItem:
      type: object
      properties:
        id: { type: integer }
        noPo: { type: string }
        namaProject: { type: string }
        customer: { type: ["string", "null"] }
        tanggalPoMasuk: { type: string }
        targetPenyelesaian: { type: ["string", "null"] }
        deadline: { type: string }
        sisaHari: { type: integer }
        picUserId: { type: ["integer", "null"] }
        picName: { type: ["string", "null"] }
        departmentId: { type: ["integer", "null"] }
        departmentName: { type: ["string", "null"] }
        status: { type: string }
        progress: { type: integer }
        catatan: { type: ["string", "null"] }
        createdByUserId: { type: ["integer", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, noPo, namaProject, deadline, sisaHari, status, progress, createdAt, updatedAt]
    PoInput:
      type: object
      properties:
        noPo: { type: string }
        namaProject: { type: string }
        customer: { type: string }
        tanggalPoMasuk: { type: string }
        targetPenyelesaian: { type: string }
        deadline: { type: string }
        picUserId: { type: integer }
        departmentId: { type: integer }
        status: { type: string }
        progress: { type: integer }
        catatan: { type: string }
      required: [noPo, namaProject, tanggalPoMasuk, deadline]
    PoUpdate:
      type: object
      properties:
        noPo: { type: string }
        namaProject: { type: string }
        customer: { type: string }
        tanggalPoMasuk: { type: string }
        targetPenyelesaian: { type: string }
        deadline: { type: string }
        picUserId: { type: integer }
        departmentId: { type: integer }
        status: { type: string }
        progress: { type: integer }
        catatan: { type: string }
    PoSummary:
      type: object
      properties:
        totalPo: { type: integer }
        poSelesai: { type: integer }
        poBelumSelesai: { type: integer }
        poDelay: { type: integer }
        poHampirDeadline: { type: integer }
        persentasePencapaian: { type: integer }
        month: { type: integer }
        year: { type: integer }
      required: [totalPo, poSelesai, poBelumSelesai, poDelay, poHampirDeadline, persentasePencapaian, month, year]