# lib/api-zod/src/generated/types/poInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Laporan Harian - Daily Work Report System API
 * OpenAPI spec version: 0.1.0
 */

export interface PoInput {
  noPo: string;
  namaProject: string;
  customer?: string;
  tanggalPoMasuk: string;
  targetPenyelesaian?: string;
  deadline: string;
  picUserId?: number;
  departmentId?: number;
  status?: string;
  progress?: number;
  catatan?: string;
}