# lib/api-zod/src/generated/types/poItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Laporan Harian - Daily Work Report System API
 * OpenAPI spec version: 0.1.0
 */

export interface PoItem {
  id: number;
  noPo: string;
  namaProject: string;
  /** @nullable */
  customer?: string | null;
  tanggalPoMasuk?: string;
  /** @nullable */
  targetPenyelesaian?: string | null;
  deadline: string;
  sisaHari: number;
  /** @nullable */
  picUserId?: number | null;
  /** @nullable */
  picName?: string | null;
  /** @nullable */
  departmentId?: number | null;
  /** @nullable */
  departmentName?: string | null;
  status: string;
  progress: number;
  /** @nullable */
  catatan?: string | null;
  /** @nullable */
  createdByUserId?: number | null;
  createdAt: string;
  updatedAt: string;
}